feat(raml_import): refactoring and parser update (AS-307)
