+++ modules/org.restlet.ext.raml/src/org/restlet/ext/raml/RamlSpecificationRestlet.java
@@ -34,7 +34,7 @@ import org.restlet.data.MediaType;
 import org.restlet.data.Method;
 import org.restlet.data.Reference;
 import org.restlet.data.Status;
-import org.restlet.ext.apispark.internal.conversion.raml.RamlTranslator;
+import org.restlet.ext.apispark.internal.conversion.raml.RamlWriter;
 import org.restlet.ext.apispark.internal.introspection.application.ApplicationIntrospector;
 import org.restlet.ext.apispark.internal.model.Definition;
 import org.restlet.representation.Representation;
@@ -197,7 +197,7 @@ public class RamlSpecificationRestlet extends Restlet {
      *         Application.
      */
     public Representation getRaml() {
-        return new StringRepresentation(new RamlEmitter().dump(RamlTranslator
+        return new StringRepresentation(new RamlEmitter().dump(RamlWriter
                 .getRaml(getDefinition())), MediaType.TEXT_PLAIN);
     }
